+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/Doc.java
@@ -5,5 +5,5 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface Doc {
-  public String msg();
+  String msg();
 }
