+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/Traversal.java
@@ -5,5 +5,5 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface Traversal {
-  public Class elementType();
+  Class elementType();
 }
